Extract the skills codec into cli/integrations/
Establish the cli/integrations/ package and move the skills artifact
codec out of cli/commands/setup.py into cli/integrations/skills.py.

setup.py has grown into a monolith mixing surface orchestration with
the codecs that render and materialize artifacts. Isolating the skills
codec is the first step of decomposing that module along its natural
seams, so each unit is named for one job.

This is a pure move with no behavior change: the moved function bodies
are byte-identical to their originals, the lazy render_skill imports
stay inside each function to preserve import-cycle behavior, and every
internal importer (setup.py, adopt.py, and the two test modules that
exercise the codec directly) is retargeted to the new module. No
re-export shim is left behind.
