+++ packages/nauro/src/nauro/cli/integrations/__init__.py
@@ -1,0 +1,1 @@
+"""Artifact codecs and orchestration policy for the setup surface."""
